+++ ase/calculators/turbomole/parameters.py
@@ -91,10 +91,6 @@ class TurbomoleParameters(dict):
             'default': None,
             'group': 'denconv',
             'key': 'denconv',
-            'mapping': {
-                'to_control': lambda a: int(-log10(a)),
-                'from_control': lambda a: 10**(-a)
-            },
             'non-define': True,
             'type': float,
             'units': None,