+++ ase/test/calculator/plumed/test_plumed.py
@@ -98,7 +98,7 @@ def test_CVs(factory):
     compares them"""
     # plumed setting
     ps = 1000 * units.fs
-    set_plumed = [f"UNITS LENGTH=A TIME={1/ps} ENERGY={units.mol/units.kJ}",
+    set_plumed = [f"UNITS LENGTH=A TIME={1 / ps} ENERGY={units.mol / units.kJ}",
                   "c1: COM ATOMS=1,2",
                   "c2: CENTER ATOMS=1,2",
                   "l: DISTANCE ATOMS=c1,c2",
@@ -218,7 +218,7 @@ def test_pbc(factory):
     traj = [atoms]
 
     ps = 1000 * units.fs
-    setup = [f"UNITS LENGTH=A TIME={1/ps} ENERGY={units.mol/units.kJ}",
+    setup = [f"UNITS LENGTH=A TIME={1 / ps} ENERGY={units.mol / units.kJ}",
              "d: DISTANCE ATOMS=1,2",
              "PRINT ARG=d STRIDE=100 FILE=COLVAR_pbc"]
 
@@ -248,7 +248,7 @@ def run(factory, inputs, name='',
 
 def setups(name=''):
     ps = 1000 * units.fs
-    set_plumed = [f"UNITS LENGTH=A TIME={1/ps} ENERGY={units.mol/units.kJ}",
+    set_plumed = [f"UNITS LENGTH=A TIME={1 / ps} ENERGY={units.mol / units.kJ}",
                   "d: DISTANCE ATOMS=1,2",
                   "FLUSH STRIDE=1",
                   f"METAD ARG=d SIGMA=0.5 HEIGHT=2 PACE=20 FILE=HILLS_{name}"]